+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 
 - [10686](https://github.com/vegaprotocol/vega/issues/10686) - Allow transfers to filter by from and to account type.
 - [10686](https://github.com/vegaprotocol/vega/issues/10764) - Network will now only trade within price monitoring bounds.
-- [](https://github.com/vegaprotocol/vega/issues/xxx)
+- [10770](https://github.com/vegaprotocol/vega/issues/10770) - Increase price monitoring triggers to 100.
+
 
 ### 🐛 Fixes